The UI only displays the link to settings when config.html exists
This plugin extends the `print_form` while it does not have a `config.html` file. From 3.3, the `print_form` method is deprecated and Moodle core checks for the presence of the file `config.html` to determine whether the plugin has settings. Adding this blank file allows for the link to be displayed.